+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog for PSFunctionTools
 
+## v0.2.0
+
++ Updated help documentation.
++ Modified `Format-FunctionName` to accept pipeline input by value.
++ Added function `Get-ParameterBlock` with an alias of `gpb`.
++ Added function `Get-FunctionAttribute` with an alias of `gfa`.
++ Added a parameter called `Detailed` to `Get-FunctionName` to write a custom object to the pipeline which includes the path to the file. Added a custom format file `psfunctionname.format.ps1xml`.
+
 ## v0.1.0
 
 + Initial files.